% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWatermark.R
\name{addWatermark}
\alias{addWatermark}
\title{plot a colour bar}
\usage{
addWatermark(txt = NULL, cex = par("cex.axis") * 0.8)
}
\arguments{
\item{txt}{text string to display, defaults to (c) MeteoSwiss <date>}

\item{cex}{character expansion factor}
}
\description{
Add watermark to plot
}
\keyword{utilities}
