% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{boxJitterPlot_std}
\alias{boxJitterPlot_std}
\title{Boxplot & jitter standardized values of continuous variables}
\usage{
boxJitterPlot_std(data, color, trans = "identity")
}
\arguments{
\item{data}{Tibble}

\item{color}{Variable passed to \code{geom_jitter()}}

\item{trans}{Character passed to \code{scale_y_continuous()}, default "identity"; "log2" suggested for outliers}
}
\value{
ggplot2 object
}
\description{
Plot boxplots and jitters of values of continuous variables which are standardized using median/mad. Ignoring dates.
Jitter uses a discrete variable that is passed through .colors.
Optional transformation for Y-asis may be used.
Implemented for plotting data from targets with many parameters; thus it ignores all \code{"color_"} variables.
Implemented for 2020-08_EKocar_fibIT.
}
