% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{brewPalFac}
\alias{brewPalFac}
\title{Brew a color palette from a factor x with optional names.}
\usage{
brewPalFac(x, n = 9, name = "OrRd", pull = NULL, namesFrom = NULL)
}
\arguments{
\item{x}{Numeric vector/factor with optional names.}

\item{n}{Number of colors for RColorBrewer (depends on the name).}

\item{name}{Name of the RColorBrewer palette.}

\item{pull}{Indices to pull colors from RColorBrewer::brewer.pal}

\item{namesFrom}{Vector of names to be added to colors; default is names(x) or x.}
}
\value{
Named factor of colors of the same length as x where the order of colors corresponds to the order of x.
}
\description{
Uses RColorBrewer palette names, default is OrRd with 9 colors
Use more/less colors (n) for a bigger/smaller range, i.e. more/less extremes at each end.
Colors are ordered according to the order of factor levels.
If namesFrom given, add names to colors; otherwise, add names(x) or x.
}
\section{Implementation}{

Levels of x are ordered as they apper in data; this is achieved by as.factor(x, levels=unique(x))
In the 2nd to the last line, we create a factor from colsn with levels ordered as they appear;
In the last line, we reorder the levels of that factor according to the order of levels in x;
xtfrm(unique(x)) reports indices of (unique) values in levels.
}

