% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCompleteCorrelationMatrices.R
\name{buildCompleteCorrelationMatrices}
\alias{buildCompleteCorrelationMatrices}
\title{buildCompleteCorrelationMatrices}
\usage{
buildCompleteCorrelationMatrices(target_dir, out_fp, proteins = FALSE,
  ncores = detectCores())
}
\arguments{
\item{target_dir}{path to directory containing files to be merged}

\item{out_fp}{path to RData file to write results to. Has to contain a substitution pattern %s}

\item{proteins}{if TRUE change names of saved objects including the word "proteins"}

\item{ncores}{number of cores to use}
}
\description{
make complete correlation matrices
}
