% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDarGRanges.R
\name{buildDarGRanges}
\alias{buildDarGRanges}
\title{buildDarGRanges}
\usage{
buildDarGRanges(path_pc_bph, path_pc_crpc, path_crpc_bph)
}
\arguments{
\item{path_pc_bph}{path to bed for pc_bph comparison}

\item{path_pc_crpc}{path to bed for crpc_pc comparison}

\item{path_crpc_bph}{path to bed for crpc_bph comparison}
}
\value{
GRanges
}
\description{
a function to build the GenomicRanges representing DARs coming from all three comparison
}
