% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_spatial_relationship}
\alias{check_spatial_relationship}
\title{tests for a spatial relationship and adds true/false to the new col and displays a progress bar}
\usage{
check_spatial_relationship(input_sf, reference_sf, sf_function, column_name)
}
\description{
tests for a spatial relationship and adds true/false to the new col and displays a progress bar
}
