% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourbar.R
\name{colourbar}
\alias{colourbar}
\title{plot a colour bar}
\usage{
colourbar(levels, colours = NULL, side = 1, units = "", ...)
}
\arguments{
\item{levels}{a vector with the levels to be used for the distinct colours or a list with components \code{levels} and \code{colours}}

\item{colours}{a vector of the colours to be used with the respective levels}

\item{side}{side at which axis labels are to be shown}

\item{units}{text for units of colourbar (at right, top end of colour bar)}

\item{...}{additional arguments passed to \code{\link{axis}}}
}
\description{
Adds a colour bar with the specified colours and levels
}
\keyword{utilities}

