% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compare_crs}
\alias{compare_crs}
\title{This function compares the crs of two sf objects}
\usage{
compare_crs(dataset1, dataset2)
}
\description{
This function compares the crs of two sf objects
}
