% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAllDarVsGeneExpressionCorrelation.R
\name{computeAllDarVsGeneExpressionCorrelations}
\alias{computeAllDarVsGeneExpressionCorrelations}
\title{computeAllDarVsGeneExpressionCorrelations}
\usage{
computeAllDarVsGeneExpressionCorrelations(genes_gr, gene_mat, mirna_mat,
  atac_mat, samples_blacklist, tmp_dir, mode = "genes", debug = FALSE,
  ncores = detectCores(), randomize = FALSE)
}
\arguments{
\item{genes_gr}{GRanges object for genes and mirna}

\item{gene_mat}{Matrix reporting expression values of protein coding genes}

\item{mirna_mat}{Matrix of expression values of mirna}

\item{atac_mat}{Matrix of accessibility levels of ATAC-seq regions}

\item{samples_blacklist}{Blacklisted samples}

\item{tmp_dir}{Path to tmp dir where store computation results}

\item{mode}{If "proteins", mirna_mat won't be used}

\item{debug}{Debug switch}

\item{ncores}{Number of cores for multiprocessing}
}
\value{
invisible
}
\description{
perform correlation analysis for each row of atac_mat against each row of gene_mat and mirna_mat. Saves one Rdata file for each atac_mat row in tmp_dir
}
\examples{

}
