% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAtacTSS.R
\name{computeAtacTSS}
\alias{computeAtacTSS}
\title{computeAtacTSS}
\usage{
computeAtacTSS(atac_tss)
}
\arguments{
\item{atac_tss}{Data.table of signals at TSS}
}
\description{
compute log2 fold change, abs med difference, p-value for TSS
}
