% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelation.R
\name{computeCorrelation}
\alias{computeCorrelation}
\title{computeCorrelation}
\usage{
computeCorrelation(method, amat, gmat, nc = detectCores(),
  debug = FALSE, verbose = FALSE)
}
\arguments{
\item{method}{correlation coefficient calculation method. See `cor` for accepted values}

\item{amat}{first matrix}

\item{gmat}{second matrix}

\item{nc}{number of cores. Defaults to all avaible cores}

\item{debug}{set debug mode. Compute correlation only for a small subset of rows of both input matrices.}

\item{verbose}{print log messages}
}
\value{
matrix
}
\description{
Compute given correlation coefficient matrix between two input matrices. Return matrix of correlations rounded to two decimal places.
}
\examples{
m1 <- matrix(rnorm(200), 20, 10, dimnames = list(NULL, letters[1:10]))
m2 <- matrix(rnorm(300), 30, 10, dimnames = list(NULL, letters[1:10]))
computeCorrelation("pearson", m1,m2)

}
