% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelationAtTSS.R
\name{computeCorrelationAtTSS}
\alias{computeCorrelationAtTSS}
\title{computeCorrelationAtTSS}
\usage{
computeCorrelationAtTSS(atac_mat, gene_mat, mirna_mat, protein_mat,
  samples_blacklist = NULL, debug = FALSE, randomize = FALSE,
  ncores = detectCores())
}
\arguments{
\item{atac_mat}{a matrix of atac signals at TSS. Gene id as rownames.}

\item{gene_mat}{gene expression matrix.}

\item{mirna_mat}{mirna expression matrix.}

\item{protein_mat}{protein expression matrix.}

\item{debug}{set debug mode. Compute correlation only for a small subset of genes.}
}
\value{
matrix with gene id as rownames and 4 columns: pearson, spearman, protein_pearson, protein_spearman
}
\description{
Compute Pearson and Spearman correlation coefficients for each gene given in atac_mat. Determines if gene is mirna or protein coding and compute correlation. If protein expression is available, compute correlation also between atac and protein levels.
}
\examples{

}
