% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelationDarsInTad.R
\name{computeCorrelationDarInTad}
\alias{computeCorrelationDarInTad}
\title{computeCorrelationDarInTad}
\usage{
computeCorrelationDarInTad(tad_gr, dars_gr, genes_gr, atac_mat, gene_mat,
  mirna_mat, proteins_mat, debug = FALSE, randomize = FALSE,
  random.seed = 123, ncores = detectCores())
}
\arguments{
\item{tad_gr}{GRanges object representing TADs}

\item{dars_gr}{GRanges object representing DARs}

\item{genes_gr}{GRanges object representing genes}

\item{atac_mat}{Data.table object representing accessibility values (ATAC-seq)}

\item{gene_mat}{matrix representing gene expression values for protein coding genes}

\item{mirna_mat}{matrix representing mirna expression values}

\item{proteins_mat}{matrix representin protein expression values}

\item{debug}{Boolean. If TRUE perform operations on small subset of TAD (3)}

\item{ncores}{Number of available cores for multithread operations}
}
\value{
matrix with gene id as rownames and 4 columns: pearson, spearman, protein_pearson, protein_spearman
}
\description{
Compute Pearson and Spearman correlation coefficients for each gene-dar couple fallingi the same topologically associated domain (TAD). First perform overlap between dar GRanges object and genes GRanges object against TAD GRanges object. For each TAD start a thread, get dars and genes overlapping that TAD, perform cartesian product and for each generated couple compute Pearson and Spearman correlation coefficients. If gene id is detected in proteins data matrix, compute correlation coefficients with that as well.
}
\examples{

}
