% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCorrelationDarsToClosestGene.R
\name{computeCorrelationDarToClosestGene}
\alias{computeCorrelationDarToClosestGene}
\title{computeCorrelationDarToClosestGene}
\usage{
computeCorrelationDarToClosestGene(dar_to_closest_gene, mirna_map,
  mirna_mat, gene_mat, atac_mat, protein_mat, samples_blacklist, id_column,
  randomize = FALSE, ncores = detectCores())
}
\arguments{
\item{dar_to_closest_gene}{data.table output from Homer}

\item{mirna_map}{mirBase to Ensembl gene id mapping}

\item{mirna_mat}{matrix of miRNA expression values}

\item{gene_mat}{matrix of genes expression values}

\item{atac_mat}{matrix of atac accessibility values}

\item{randomize}{flag. If true shuffle samples order prior to correlation computation}

\item{proteins_mat}{matrix of protein expression values}
}
\description{
compute correlation dar to closest gene according to Homer annotation
}
