% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDarInTadCorrelationV3.R
\name{computeDarInTadCorrelationV3}
\alias{computeDarInTadCorrelationV3}
\title{computeDarInTadCorrelationV3}
\usage{
computeDarInTadCorrelationV3(df, atac_mat, mirna_mat, gene_mat,
  protein_mat, samples_blacklist, id_name, randomize = FALSE,
  ncores = detectCores())
}
\arguments{
\item{df}{the table to process}

\item{mirna_mat}{matrix of expression values of miRNA}

\item{gene_mat}{matrix of expression values of genes}

\item{protein_mat}{matrix of expression values of proteins}

\item{randomize}{flag. If true shuffle sample order prior of correlation computation}
}
\description{
compute correlation from processed Betools intersect output file
}
