% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_note.R
\name{create_note}
\alias{create_note}
\title{Initializes files and directories for a new R project}
\usage{
create_note(content = "")
}
\arguments{
\item{content}{Content to be written, default None}
}
\value{
None
}
\description{
This function just creates a new note with a timestamp
}
\examples{
# create_note()

}
