% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_py_project.R
\name{create_py_project}
\alias{create_py_project}
\title{Initializes files and directories for a new Python project}
\usage{
create_py_project(path_to_templates = "/Users/phil/Documents/templates")
}
\arguments{
\item{path_to_templates}{The path to the location of the template, css, and
other files.}
}
\value{
base directories and Python scripts for a new project.
}
\description{
This function just creates a files and directories I use regularly in a
new Python project.
}
\examples{
# create_py_project(
#   path_to_templates = '/my_path_to_templates'
# )

}
