% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_project.R
\name{create_r_project}
\alias{create_r_project}
\title{Initializes files and directories for a new R project}
\usage{
create_r_project(path_to_templates = "/Users/phil/Documents/templates")
}
\arguments{
\item{path_to_templates}{The path to the location of the template, css, and
other files.}
}
\value{
base directories and R scripts for a new project.
}
\description{
This function just creates a files and directories I use regularly in a
new R project.
}
\examples{
# create_r_project(
#   path_to_templates = '/my_path_to_templates'
# )

}
