% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_script}
\alias{create_script}
\title{Creates R script template in working directory}
\usage{
create_script(
  filename = NULL,
  draftname = "_extensions/myhelpers/template.R",
  ext_name = "myhelpers"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have to
add \code{.R}.}

\item{draftname}{The path to the draft file. If not indicated, the package
included draft file will be used.}

\item{ext_name}{The name of extension to get the files from. Default
"myhelpers".}
}
\value{
R script based on the template.
}
\description{
This function just creates a R script file based on the saved template as
 draft in the package folder.
}
\examples{
# create_script(
#   filename = NULL,
#   draftname = '_extensions/myhelpers/template.R',
#   ext_name = 'myhelpers'
# )

}
