% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{directory}
\alias{directory}
\title{This function reads a csv containing download links for swissalti3d and downloads the .tif files}
\format{
An object of class \code{character} of length 1.
}
\usage{
directory
}
\description{
This function reads a csv containing download links for swissalti3d and downloads the .tif files
}
\keyword{datasets}
