% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{discretize_namedIntervals}
\alias{discretize_namedIntervals}
\title{Convert X to discretized values with elements in the form of intervals min-max.}
\usage{
discretize_namedIntervals(X, ...)
}
\arguments{
\item{X}{Numeric vector or list}

\item{disc}{Default "equalfreq"}

\item{nbins}{Default NROW(X)^(1/3)}
}
\description{
See infotheo::discretize for parameters; defaults are:
}
\section{Old implementation}{

discretize_namedIntervals_vector <- function(X,...) {
## See infotheo::discretize for parameters; defaults are: disc="equalfreq", nbins=sqrt(NROW(x)).
## Returns discretized data in the form of intervals min-max.
require(infotheo)
d<- infotheo::discretize(X,...)$X
m<-tapply(X,d,min)
M<-tapply(X,d,max)
d_int <- paste0(format(m), "-", format(M))
names(d_int) <- names(m)
d_int\link{m==M} <- format(m)\link{m==M}
#d_int\link{d}
browser()
#as.character((sapply(as.character(d), function(cd) {d_int\link{cd}})))
as.character(d_int\link{as.character(d)})
}
}

