% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{discretizeTime_equalInterval}
\alias{discretizeTime_equalInterval}
\title{Get discretized time intervals in form "H:MM-H:MM"}
\usage{
discretizeTime_equalInterval(times, time0, nbins)
}
\arguments{
\item{times}{Vector of times in POSIX* format}

\item{time0}{???}

\item{nbins}{Number of intervals}
}
\value{
Character vector of discretized time intervals in form "H:MM-H:MM"
}
\description{
Get discretized time intervals in form "H:MM-H:MM"
}
