% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{download_and_merge_DEM}
\alias{download_and_merge_DEM}
\title{This function combines the three functions used to generate download links for swissalti3d, downloading the files and mergin the .tifs}
\usage{
download_and_merge_DEM(region, directory, resolution, name)
}
\description{
This function combines the three functions used to generate download links for swissalti3d, downloading the files and mergin the .tifs
}
