% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpFiles.R
\name{dumpFiles}
\alias{dumpFiles}
\title{dumpFiles}
\usage{
dumpFiles(target_dir, out_fp)
}
\arguments{
\item{out_fp}{path to result files (should have 2 placeholders %s)}

\item{targer_dir}{path to directory holding single Rd files}
}
\description{
Write matrix files from files reporting correlations
}
