% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandNA.R
\name{expandNA}
\alias{expandNA}
\alias{shrink}
\alias{shrink}
\title{Expand arrays with missing values and subset arrays}
\usage{
expandNA(x, dims)

shrink(x, dims)
}
\arguments{
\item{x}{array with input data}

\item{dims}{vector with dimensions describing the array extent (dimension \code{i} of output ranges from \code{1:dims[i]})}
}
\value{
Output array of dimension \code{dims}
}
\description{
This function expands a multi-dimensional array to the dimensionality indicated by filling in missing values. No new dimensions are created but existing dimensions are enlarged. The algorithm is considerably faster if only the last dimension is changed.
}
\details{
Expansion is performed such that the original data in \code{x} sits in the lower left corner of the array (i.e. for each dimension the indices \code{1:dim(x)}). Correspondingly, shrinking is performed by selected the lower-left corner as a subset.
}
\examples{
tt <- outer(1:4, 1:3, paste, sep=',')
expandNA(tt, c(5,5))
shrink(tt, c(2,2))

}
