% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filled_contour.R
\name{filled_contour}
\alias{filled_contour}
\title{Plot Filled Contours}
\usage{
filled_contour(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out
  = ncol(z)), z, levels = pretty(zlim, 12), col = NULL, xlim = range(x),
  ylim = range(y), zlim = range(z[is.finite(z)]), add = FALSE,
  axes = TRUE, type = c("contour", "image"), col.axis = 1, col.box = 1,
  add.boundary = TRUE, col.contour = "grey", lty.contour = 1,
  lwd.contour = 1, drawlabels = FALSE, ...)
}
\arguments{
\item{x,y}{Locations of grid cells for values of \code{z}.}

\item{z}{Matrix containing the values to be plotted.}

\item{levels}{Numeric vector of levels at which to draw contours.}

\item{col}{Vector of colours to use for shading in between contours.}

\item{xlim,ylim,zlim}{x- y- and z-limits for the plot.}

\item{add}{Logical, hsould plot be added to existing plot?}

\item{axes}{Logical, should axes and bounding box be drawn?}

\item{type}{One of 'contour' or 'image' for contour and image plots respectively.}

\item{col.axis,col.box}{Colour for axes and bounding box.}

\item{add.boundary}{Logical, should filled contours be separated with a
boundary line (drawn by \code{\link[graphics]{contour}})?}

\item{col.contour,lty.contour,lwd.contour}{Colour, line type, and line width
for boundary contours.}

\item{drawlabels}{Logical, contours are labelled if \code{TRUE}.}

\item{...}{Additional arguments to \code{\link{plot.window}},
  \code{\link{title}}, \code{\link{Axis}} and \code{\link{box}}.}
}
\description{
Contrary to \code{\link[graphics]{filled.contour}}, this function provides
the atmoic function for contour shading in analogy with
\code{\link[graphics]{image}} and \code{\link[graphics]{contour}}.
}
\examples{
x <- seq(1,100)
y <- seq(1,80)
z <- outer(x,y, function(x,y) sin(x/10) + cos(y / 50))

# contour plot
filled_contour(x,y,z)
}
\keyword{plot}

