% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_swissalti3d_urls}
\alias{generate_swissalti3d_urls}
\title{This function generates download links from the swissalti3d product of swisstopo for an input region}
\usage{
generate_swissalti3d_urls(region, directory, resolution)
}
\description{
This function generates download links from the swissalti3d product of swisstopo for an input region
}
