% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationTable.R
\name{getAnnotationTable}
\alias{getAnnotationTable}
\title{load custom gff/gtf annotaion}
\usage{

  getAnnotationTable(annotation_table_path = "/bmt-data/genomics/reference/Homo_sapiens.GRCh38.90.chr.gtf",
  chromosome_length_path = "/bmt-data/genomics/reference/GRCh38p11_chr_lengths.tsv",
  annotation_style = "UCSC", assembly = "hg38",
  include_chromosomes = FALSE, include_seqlengths = TRUE)
}
\arguments{
\item{annotation_table_path}{a path to gtf file}

\item{chromosome_length_path}{a path to chromosome lengths}

\item{annotation_style}{annotation style}

\item{assembly}{genome version}

\item{include_chromosomes}{remove chromosomal ranges from returned object}

\item{include_seqlengths}{add seqingfo with chromosome lengths}
}
\value{
GenomicRanges object
}
\description{
load custom gff/gtf annotaion
}
\examples{
getAnnotationTable()

}
