% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMoltenCorrelationMatrix.R
\name{getMoltenCorrelationMatrix}
\alias{getMoltenCorrelationMatrix}
\title{getMoltenCorrelationMatrix}
\usage{
getMoltenCorrelationMatrix(m1, m2, varnames, ncores = detectCores(),
  debug = FALSE, verbose = FALSE)
}
\arguments{
\item{m1}{first matrix}

\item{m2}{second matrix}

\item{varnames}{character vector. Names to assign to molten columns. First varname should describe m1, second describe m2}

\item{debug}{set debug mode. Compute correlation only for a small subset of rows of both input matrices.}
}
\value{
data.table
}
\description{
Compute Pearson and Spearman correlation matrices between the two input matrices. Melt each and perform an inner join. Return the molten joined matrix.
}
\examples{

}
