% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{ggBinary}
\alias{ggBinary}
\title{Logistic Regression boxplot}
\usage{
ggBinary(data1, X, Y, color, xlab, ylab, notch = FALSE, ...)
}
\description{
Logistic Regression boxplot
}
\section{Old implementation}{

plotBinary <- function (X, Y, ...) {
plot(X, jitter(Y, factor = 0.1), col = rgb(0, 0, 0, 0.5), pch = 19, ylim = c(-0.2, 1.2), ...)
boxplot(X ~ Y, horizontal = TRUE, notch = TRUE, add = TRUE,
at = c(-0.1, 1.1), width = c(0.1, 0.1), col = "grey",
boxwex = 0.1, yaxt = "n")
}
}

