% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGr.R
\name{importGr}
\alias{importGr}
\title{Import path as GenomicRange. Subtract one to starting position.}
\usage{
importGr(path)
}
\arguments{
\item{path}{path to bed file.}
}
\value{
gr a GenomicRange object
}
\description{
Import path as GenomicRange. Subtract one to starting position.
}
\examples{
importGr("some/path/to/file.bed")

}
