% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHomer.R
\name{importHomer}
\alias{importHomer}
\title{importHomer}
\usage{
importHomer(path, comparison = NULL, id_name = "dar_id")
}
\arguments{
\item{path}{path to the annotation table}

\item{comparison}{comparison name}

\item{id_name}{name to assign to the first column}
}
\description{
a function to load Homer annotation table as data.table
}
