% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMappingTable.R
\name{importMappingTable}
\alias{importMappingTable}
\title{Import path as data.table. Rename columns to uniprot and ensgid and collapse by uniprot id. Generate 1 to many relationship table}
\usage{
importMappingTable(path, selected_columns = NULL)
}
\arguments{
\item{path}{path to table.}

\item{selected_columns}{columns to extract from path.}
}
\value{
a data.table
}
\description{
Import path as data.table. Rename columns to uniprot and ensgid and collapse by uniprot id. Generate 1 to many relationship table
}
\examples{
importMappingTable("some/path/to/mapping.txt", selected_columns = c(1,2))
importMappingTable("some/path/to/mapping.txt")
}
