% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{join_geodata}
\alias{join_geodata}
\title{This function joins data from two sf objects}
\usage{
join_geodata(base_data, join_data, base_key, join_key, attributes = NULL)
}
\description{
This function joins data from two sf objects
}
