\name{load_beyonce_bayes}
\alias{load_beyonce_bayes}
\title{Changes ggplot style to specific beyonce style and \code{bayesplot}-functions.}
\description{
This function just utilize beyonce colors into the ggplot style. It has predefined values, which can be altered. Furthermore, it alters \code{bayesplot} functions values for plots.
}
\usage{
load_beyonce_bayes(
    pal = 126,
    textcolor = 5,
    ticksgrid = 4,
    legendtext = 1
    )
}
\arguments{
  \item{pal}{Name here the palette you want to utilize in the setting.}
  \item{textcolor}{This variable defines textcolor and background.}
  \item{ticksgrid}{This color defines ticks and grids.}
  \item{legendtext}{This color defines legendtext.}
}
\examples{
# load_beyonce_bayes()
}
