% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_colorblind.R
\name{load_colorblind}
\alias{load_colorblind}
\title{Colorblind-friendly palette.}
\usage{
load_colorblind()
}
\value{
A Colorblind-friendly palette in character vector \code{cbp}.
}
\description{
This function loads a colorblind palette into the environment
}
\examples{
# load_colorblind()

}
