\name{load_grey}
\alias{load_grey}
\title{Changes ggplot-theme to grey with x-major grid. }
\description{
With this function you alter the ggplot-theme to grey with x-major grid. Furthermore, font is adjusted. If you haven't installed \code{Fira Sans} change value. Textcolor can be adjusted by \code{color}.
}
\usage{
load_grey(
    font = "Fira Sans",
    color = "white"
)
}
\arguments{
  \item{font}{Define the font you want.}
  \item{color}{Define text color.}
}
\examples{
# load_grey(font = "Fira Sans", color = "white")
}
