% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_slide_colors.R
\name{load_slide_colors}
\alias{load_slide_colors}
\title{Adjusted theme for slides}
\usage{
load_slide_colors()
}
\value{
Set ggplot-Theme to preferred color palette as indicated by the
 function.
}
\description{
Changes ggplot style to specific style concerning slides color setting.
}
\examples{
# load_slide_colors()

}
