% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_slidecols.R
\name{load_slidecols}
\alias{load_slidecols}
\title{Adjusted theme for slides}
\usage{
load_slidecols()
}
\value{
Set ggplot-Theme to preferred color palette as indicated by the
 function.
}
\description{
Changes ggplot style to specific style concerning slides color setting.
}
\examples{
# load_slidecols()

}
