% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Draw geographical maps}
\usage{
map(database = "world", interior = TRUE, regions = ".", add = FALSE,
  ...)
}
\arguments{
\item{database}{map database to use see \code{\link[maps]{map}}}

\item{interior}{If \code{FALSE}, interior segments are not drawn.
An interior segment is a line segment of the map that bounds two
of the polygons to be drawn. This argument is ignored if fill is \code{TRUE}.}

\item{regions}{Specific region to plot  see \code{\link[maps]{map}}}

\item{add}{logical, should map be added to existing plot?}

\item{...}{parameters passed to \code{\link[maps]{map}} from the \code{maps} package}
}
\description{
Fix the interior polygons in the standard map
}
\seealso{
\code{\link[maps]{map}} from the \code{maps} package
}
\keyword{utilities}

