% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{merge_raster_files}
\alias{merge_raster_files}
\title{This function merges all .tif files in a directory}
\usage{
merge_raster_files(directory, name, resolution)
}
\description{
This function merges all .tif files in a directory
}
