% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeGtrd.R
\name{mergeGtrd}
\alias{mergeGtrd}
\title{mergeGtrd}
\usage{
mergeGtrd(gr, gtrd)
}
\arguments{
\item{gr}{ranges to search TFBS for}

\item{gtrd}{GTRD GenomicRanges}
}
\value{
data.table
}
\description{
Find TFBS overalpping features
}
\examples{
mergeGtrd(dt, gtrd)
}
