% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeGtrdFromDataFrame.R
\name{mergeGtrdFromDataFrame}
\alias{mergeGtrdFromDataFrame}
\title{MergeGtrdFromDataFrame}
\usage{
mergeGtrdFromDataFrame(input_df, gid_df, gtrd, by)
}
\arguments{
\item{input_df}{data.frame to be merged}

\item{gid_df}{data frame of coordinates to look for TFBS}

\item{gtrd}{GenomicRanges object of TFBS coordinates}
}
\value{
data.table
}
\description{
MergeGtrdFromDataFrame
}
\examples{

}
