% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeWithEnsembl.R
\name{mergeWithEnsembl}
\alias{mergeWithEnsembl}
\title{mergeWithEnsembl}
\usage{
mergeWithEnsembl(cormat, ann.gene, mirna_map, target_column = NULL)
}
\arguments{
\item{cormat}{target data.table}

\item{ann.gene}{Ensembl annotation data.frame}

\item{mirna_map}{mapping between mirBase id and Ensembl gene id}
}
\value{
merged data.table
}
\description{
mergeWithEnsembl
}
