% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{multiEffectPlot}
\alias{multiEffectPlot}
\title{Logistic Regression plot of effects}
\usage{
multiEffectPlot(
  model,
  effectNames = c("anova", "summary"),
  OR = FALSE,
  test = "Chi"
)
}
\description{
Developed for OBV 2019-10-13
}
