\name{my_pause}
\alias{my_pause}
\title{Pause and garbage colletion}
\description{
Just a helper function to pause system and then do garbage collection (helpful when running a script with intense RAM use).
}
\usage{
my_pause(
  time = 30
)
}
\arguments{
  \item{time}{Indicate the pause in seconds}
}
\examples{
#my_pause(
#    time = 30
#)
}
