% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\docType{package}
\name{myHelpers}
\alias{myHelpers}
\title{myHelpers: Misc helper functions for R}
\description{
Misc helper functions for R
}
\section{Base}{

setRowNames
}

\section{STAT + PLOTS}{

boxJitterPlot_std
write_boxplot_descStat
plot_corr
}

\section{Write}{

write.table.rowNames
}

\section{Colors}{

brewPalFac
brewPalCont
}

\section{Conversions}{

defactorChr
defactor
discretize_namedIntervals
discretizeTime_equalInterval
capwords
}

\section{Plots}{

MDScols
MASS_MDScols (DEPRICATED)
plotIsoMDS
multiEffectPlot
ggBinary
}

\section{Calculations}{

corDist
scaleRobust
}

\author{
Peter Juvan, \email{peter.juvan@gmail.com}
}
