% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nc_time.R
\name{nc_time}
\alias{nc_time}
\title{nc_time}
\usage{
nc_time(nc)
}
\arguments{
\item{nc}{netcdf file connection}
}
\description{
read netcdf times and convert to RDates
}
\seealso{
\code{\link{nc_open}} for opening netcdf files,
  \code{\link{as.Date}} and \code{\link{as.POSIXct}} for information on R
  Date format
}
\keyword{utils}

