% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_write.R
\name{nc_write}
\alias{nc_write}
\title{nc_write}
\usage{
nc_write(nctempfile, file, varname, data, append = FALSE, ...)
}
\arguments{
\item{nctempfile}{path to netcdf template (see details)}

\item{file}{path to output netcdf file}

\item{varname}{variable name to write output to (if set to \code{NULL},
the first variable in the template file will be used)}

\item{data}{array (or vector) containing the data to be written to the output
file}

\item{append}{logical, should data be appended to existing netcdf file?}

\item{...}{additional parameters passed to \code{\link{ncvar_put}}}
}
\description{
Quickly write content to netcdf following a netcdf template file
}
\details{
This function copies an already existing NetCDF file to the new
location and writes data to the variable indicated by \code{varname}. The
dimensions and all other attributes are copied from \code{nctempfile}.
}
\note{
The climate data operators \code{cdo} command line tool is used to
produce an empty netcdf file
}

