% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayes_acf.R
\name{plot_bayes_acf}
\alias{plot_bayes_acf}
\alias{create_bayes_acf}
\title{Plots diagnostics: posterior autocorrelation.}
\usage{
plot_bayes_acf(modelfit, nottoplot = 3, neachplot = 5, folder, model)

create_bayes_acf(modelfit, nottoplot = 3, neachplot = 5, folder, model)
}
\arguments{
\item{modelfit}{Object as result of fit from \code{brms}.}

\item{nottoplot}{Indicate if last diagnostic parameters should not be
printed by putting in an integer.}

\item{neachplot}{Indicate how many variables are plotted in a plot.}

\item{folder}{Indicate folder where to save the plot (from working
directory).}

\item{model}{Indicate a model name to easier identify the png-file later.}
}
\value{
Saves plots for diagnostics (autocorrelation).
}
\description{
Creates diagnostic plots of bayesian estimation (\code{brms} object) of
 autocorrelation. Automatically saves png-files.
}
\examples{
# plot_bayes_acf(
#   modelfit,
#   nottoplot = 3,
#   neachplot = 5,
#   folder = './output/',
#   model = 'fitm1'
# )

}
