% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayes_combo.R
\name{plot_bayes_combo}
\alias{plot_bayes_combo}
\alias{create_bayes_combo}
\title{Plots diagnostics: posterior density and trace}
\usage{
plot_bayes_combo(modelfit, nottoplot = 3, neachplot = 5, folder, model)

create_bayes_combo(modelfit, nottoplot = 3, neachplot = 5, folder, model)
}
\arguments{
\item{modelfit}{Object as result of fit from \code{brms}.}

\item{nottoplot}{Indicate if last diagnostic parameters should not be
printed by putting in an integer.}

\item{neachplot}{Indicate how many variables are plotted in a plot.}

\item{folder}{Indicate folder where to save the plot (from working
directory).}

\item{model}{Indicate a model name to easier identify the png-file later.}
}
\value{
Saves plots for diagnostics (density and trace).
}
\description{
Creates diagnostic of Bayesian estimation (\code{brms} object) of
 posterior density and trace.
}
\examples{
# plot_bayes_acf(
#   modelfit,
#   nottoplot = 3,
#   neachplot = 5,
#   folder = './output/',
#   model = 'fitm1'
# )

}
