% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plots posterior distribution}
\usage{
plot_posterior(
  fit,
  ndraws = 150,
  type = "ecdf_overlay",
  font = "Fira Sans",
  model = NULL,
  folder = "./"
)
}
\arguments{
\item{fit}{A \code{brms} fit to plot.}

\item{ndraws}{Set the draws for plotting. \code{default} is 150.}

\item{type}{Choose the type for plotting, for help check \code{??pp_check}.
Example \code{bars} or \code{dens_overlay}.}

\item{font}{Set the font for text in the plot.}

\item{model}{Indicate the model for identification later. Default is
\code{NULL}.}

\item{folder}{Indicate a different folder than working directory for saving.}
}
\value{
Saves plots for diagnostics (autocorrelation).
}
\description{
Creates a plot with function \code{pp_check} from \code{bayesplot} and
 saves it directly.
}
\examples{
# plot_posterior(
#   fit,
#   ndraws = 150,
#   type = 'ecdf_overlay',
#   font = 'Fira Sans',
#   model = NULL,
#   folder = './'
# )

}
