% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raincloud.R
\name{plot_raincloud}
\alias{plot_raincloud}
\alias{create_raincloud}
\title{Plots a raincloud (metric variable)}
\usage{
plot_raincloud(df, metricvar)

create_raincloud(df, metricvar)
}
\arguments{
\item{df}{A Data frame.}

\item{metricvar}{A metric variable within that data frame.}
}
\value{
Returns a raincloud of the metric variable.
}
\description{
This function creates a simple raincloud of a single metric variable. It
 utilizes \code{beyonce} for coloring, \code{ggdist} and \code{gghalves}
 for additional plotting.
}
\examples{
# plot_raincloud(
#   df,
#   metricvar
# )

}
