% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\title{plot skill scores from data frame}
\usage{
plot_scores(df, value.var = "value",
  dpath = "/store/msclim/bhendj/EUPORIAS", type = c("image", "contour"),
  levels = NULL, colours = NULL, na.colour = NULL, add.map = TRUE,
  map.interior = FALSE, database = "world", title = "", ...)
}
\arguments{
\item{df}{data frame with one score to plot (one instance)}

\item{value.var}{variable name to be displayed}

\item{dpath}{parent directory for grid files}

\item{type}{user defined type for colour shading. Either individual grid boxes (\code{image})
or shaded contours (\code{contour}) are plotted.}

\item{levels}{breaks for colour scale}

\item{colours}{user defined colours (one less than levels)}

\item{na.colour}{user defined colour for missing values (avoided when \code{NULL})}

\item{add.map}{map is added if \code{TRUE}}

\item{map.interior}{should country boundaries be plotted?}

\item{database}{name of map database to be used}

\item{title}{text to be displayed in topleft corner of plot}

\item{...}{additional arguments passed to \code{\link[graphics]{image}}}
}
\description{
Draw map of skill scores stored in a data frame. Incomplete rows and
columns are expanded (e.g. missing values, land sea mask, etc.)
}
\keyword{util}

