% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{plotIsoMDS}
\alias{plotIsoMDS}
\title{Plot MDS that replaces limma::plotMDS}
\usage{
plotIsoMDS(
  data9,
  scale = F,
  center = F,
  labels = names(data9),
  col = NULL,
  cex = 1,
  main = NULL,
  cex.main = 1,
  xlab = "Coordinate 1",
  ylab = "Coordinate 2",
  ...
)
}
\arguments{
\item{main}{If TRUE, title is generated automatically; default NULL, string othervise}
}
\description{
Uses isoIMS
Similar to limma::plotMDS, except that is uses all parameters for distance calculation,
while limma uses only top=XX genes, but XX cannot be set
}
