% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_forecast.R
\name{read_forecast}
\alias{read_forecast}
\title{Read Seasonal Forecasts from Repository}
\usage{
read_forecast(model = "ecmwf-system4", index = "tas", grid = "global2",
  method = "none", init = "05", lon = 8.55, lat = 47.37, loi = NULL,
  lai = NULL, granularity = "monthly", source = c("euporias",
  "operational"))
}
\arguments{
\item{model}{Name of forecasting system}

\item{index}{Name of index (variable) to be read}

\item{grid}{Name of grid to be read}

\item{method}{Method string for debiasing of forecast}

\item{init}{Month of initialization (e.g. '05')}

\item{lon}{vector with the longitude to be read in}

\item{lat}{vector with the latitude to be read in}

\item{loi}{vector with grid index to be read in (alternative to \code{lon})}

\item{lai}{vector with grid index to be read in (alternative to \code{lat})}

\item{granularity}{granularity of forecast data to be read}

\item{source}{where should forecasts be retrieved from (see details)?}
}
\description{
Similar to \code{\link{read_scores}}, \code{read_forecast} reads forecasts
from the repository.
}
\details{
Alternatives are for the forecasts to be retrieved from the research
  directory (\code{source = 'euporias'}) or from the direcotry with
  operational forecasts (\code{source = 'operational'}). In the latter case,
  the grid, method, and granularity arguments are void as these will always
  be set to native, none and monthly respectively.
}
\keyword{util}

