% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_forecasts.R
\name{read_forecasts}
\alias{get_fcfiles}
\alias{read_forecasts}
\title{Read Seasonal Forecasts from Repository}
\usage{
read_forecasts(fcfiles, index = NA, lon = NULL, lat = NULL, loi = NULL,
  lai = NULL, expand = TRUE)

get_fcfiles(model = "ecmwf-system4", grid = "global2", index = "tas",
  method = "none", init = "05", granularity = "monthly",
  source = c("euporias", "operational"))
}
\arguments{
\item{fcfiles}{a vector of file paths to the forecast file collection}

\item{index}{name of variable (index) to be read}

\item{lon,lat}{the longitude and latitude to be read for point forecasts,
or the longitude and latitude limits for a subset of the forecast domain}

\item{loi,lai}{the grid indices to be read (alternative to \code{lon, lat})}

\item{expand}{logical, should forecast arrays be expanded if differing
  numbers of ensemble members are available in hindcast set?}

\item{model}{Name of forecasting system}

\item{grid}{Name of grid to be read}

\item{method}{Name of method for debiasing of forecast}

\item{init}{Month of initialization (e.g. '05')}

\item{granularity}{granularity of forecast data to be read}

\item{source}{where should forecasts be retrieved from (see details)?}
}
\description{
\code{read_forecasts} reads and collates (a slice of) forecasts.
\code{get_fcfiles} assembles the available forecast files for a collection of
forecasts (e.g. for a specific index, grid, granularity, and bias correction
method).
}
\details{
Alternatives are for the forecasts to be retrieved from the research
  directory (\code{source = 'euporias'}) or from the direcotry with
  operational forecasts (\code{source = 'operational'}). In the latter case,
  the grid, method, and granularity arguments are void as these will always
  be set to native, none and monthly respectively.
}
\note{
Reading in subsets only works for continous subset not wrapping
around the domain. That is, for data on a global grid with positive
longitudes, a European domain cannot be extracted.
}
\keyword{util}

