% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_ncdf.R
\name{read_ncdf}
\alias{get_fcfiles}
\alias{read_ncdf}
\alias{read_single}
\title{Read Data from a NetCDF}
\usage{
read_ncdf(x, expand = FALSE, ...)

read_single(x, index = NA, lon = NULL, lat = NULL, loi = NULL,
  lai = NULL, mask = NULL, tlim = NULL, ti = NULL, time = NULL)

get_fcfiles(model = "ecmwf-system4", grid = "global2", index = "tas",
  method = "none", init = "05", granularity = "monthly",
  source = c("euporias", "operational", "demo"))
}
\arguments{
\item{x}{a vector of file paths to the forecast file collection}

\item{expand}{logical, should forecast arrays be expanded if differing
numbers of ensemble members are available in hindcast set?}

\item{...}{parameters passed on to \code{\link{read_single}}}

\item{index}{Variable name to be read}

\item{lon,lat}{the longitude and latitude to be read for point forecasts,
or the longitude and latitude limits for a subset of the forecast domain}

\item{loi,lai}{the grid indices to be read (alternative to \code{lon, lat})}

\item{mask}{logical vector with grid point indices to be read (see details).}

\item{tlim}{vector of start and end dates to be read.}

\item{ti}{time indices to be read}

\item{time}{vector of dates to be read. The comparison is performed on
on the same format as \code{time} (i.e. if \code{is.Date(time)}, the
comparison is done transforming the ncdf time to \code{as.Date})}

\item{model}{Name of forecasting system}

\item{grid}{Name of grid to be read}

\item{method}{Name of method for debiasing of forecast}

\item{init}{Month of initialization (e.g. '05')}

\item{granularity}{granularity of forecast data to be read}

\item{source}{where should forecasts be retrieved from (see details)?}
}
\description{
\code{read_ncdf} reads and collates (a slice of) data from a
  single netcdf file or a collection of netcdf files. \code{read_single} is
  the underlying function that does the heavy lifting on the single files
  (not exported, for programming purposes only). \code{get_fcfiles} assembles
  the available forecast files for a collection of forecasts (e.g. for a
  specific index, grid, granularity, and bias correction method). Reading
  subsets of NetCDF files allows you to subset the file by longitude and
  latitude, by grid index, by time, or as a compact representation of a
  selection of grid points (e.g. land only).
}
\details{
Alternatives are for the forecasts to be retrieved from the research
  directory (\code{source = 'euporias'}) or from the direcotry with
  operational forecasts (\code{source = 'operational'}). In the latter case,
  the grid, method, and granularity arguments are void as these will always
  be set to native, none and monthly respectively.
}
\note{
Reading in subsets only works for continous subset not wrapping
around the domain. That is, for data on a global grid with positive
longitudes, a European domain cannot be extracted.
}
\seealso{
read_single
}
\keyword{util}

