% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_scores.R
\name{read_scores}
\alias{read_scores}
\title{read_scores}
\usage{
read_scores(scores = c("EnsCorr", "FairRpss"), indices = "tas",
  methods = "none", initmonths = "05", grids = "global2",
  granularities = "seasonal", ccrs = c(FALSE, TRUE), detrends = FALSE,
  leads = 2, dpath = "/store/msclim/bhendj/EUPORIAS")
}
\arguments{
\item{scores}{character vector of skill scores to be read}

\item{indices}{character vector of variable names}

\item{methods}{character vector of bias correction methods}

\item{initmonths}{character vector of initialization months}

\item{grids}{character vector of grids}

\item{granularities}{character vector of temporal resolutions}

\item{ccrs}{logical should recalibrated indices be used?}

\item{detrends}{logical should detrended series be used?}

\item{leads}{index of lead times}

\item{dpath}{head of directory tree}
}
\description{
Read skill scores to data frame
}
\keyword{util}

