% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/removeOverlappingDarsBySpearmanCorrelationValue.R
\name{removeOverlappingDarsByCorrelationValue}
\alias{removeOverlappingDarsByCorrelationValue}
\title{removeOverlappingDarsBySpearmanCorrelationValue}
\usage{
removeOverlappingDarsByCorrelationValue(df, comparison)
}
\arguments{
\item{df}{complete correlation matrix}

\item{comparison}{string representing one of the three comparisons in the analysis (pc_bph, crpc_pc, crpc_bph)}
}
\value{
fitered cormat
}
\description{
Remove overlapping DARs correlating with the same gene by absolute Spearman correlation value
}
\examples{

}
