% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{setRowNames}
\alias{setRowNames}
\title{Equivalent to setNames(object) for arrays}
\usage{
setRowNames(x, rowNames)
}
\arguments{
\item{x}{A matrix, array or data frame}

\item{rowNames}{New names for rows}
}
\value{
Input x with new row names
}
\description{
Sets the first component of dimnames(x)
}
\section{Implementation}{

Alternative way, does not work for data.frame:
\code{function(x, rowNames) return(structure(x, dimnames=c(list(rowNames), dimnames(x)[-1])))}
}

\examples{
setRowNames(matrix(1:4, nrow=2, ncol=2), c("a","b"))
setRowNames(array(1:8, dim=c(2,2,2)), c("a","b"))
\dontrun{setRowNames(1:2, c("a","b")) # Error 'dimnames' applied to non-array}
}
