% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDarsInTadCorrelationTable.R
\name{subsetDarsInTadCorrelationTable}
\alias{subsetDarsInTadCorrelationTable}
\title{subsetDarsInTadCorrelationTable}
\usage{
subsetDarsInTadCorrelationTable(tad_gr, dars_gr, genes_gr, pearson_cormat,
  pearson_pvalue, spearman_cormat, spearman_pvalue, pearson_protein_cormat,
  pearson_protein_pvalue, spearman_protein_cormat, spearman_protein_pvalue,
  debug = FALSE, ncores = detectCores())
}
\arguments{
\item{tad_gr}{GRanges representing TADs}

\item{dars_gr}{GRanges represeting dars or peaks}

\item{genes_gr}{GRanges representing genes}

\item{pearson_cormat}{Matrix of Pearson correlation coefficients between genes and atac features}

\item{pearson_pvalue}{Matrix of p-values associated with Pearson correlation coefficients between genes and atac features}

\item{spearman_cormat}{Matrix of Spearman correlation coefficients between genes and atac features}

\item{spearman_pvalue}{Matrix of p-values associated with Spearman correlation coefficients between genes and atac features}

\item{pearson_protein_pvalue}{Matrix of p-values associated with Pearson correlation coefficients between proteins and atac features}

\item{spearman_protein_pvalue}{Matrix of p-values associated with Spearman correlation coefficients between proteins and atac features}

\item{pearson_protein}{Matrix of Pearson correlation coefficients between proteins and atac features}

\item{spearman_protein}{Matrix of Spearman correlation coefficients between proteins and atac features}
}
\value{
a data.frame
}
\description{
Perform the subsetting of correlation tables by finding dars in tad.
}
