% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSD.R
\name{twoSD}
\alias{twoSD}
\title{Standardize variables}
\usage{
twoSD(x)
}
\arguments{
\item{x}{Vector to scale.}
}
\value{
Returns a scaled vector by two SD and centered.
}
\description{
Scales variables by two SD as recommended by Gelman.
}
\examples{
# twoSD(x)

}
