% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{write.table.rowNames}
\alias{write.table.rowNames}
\title{Write a table with rownames using write.table}
\usage{
write.table.rowNames(
  x,
  col1name = "rowNames",
  quote = F,
  sep = "\\t",
  row.names = F,
  col.names = T,
  ...
)
}
\description{
Write a table with rownames using write.table
}
