% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{write_boxplot_descStat}
\alias{write_boxplot_descStat}
\title{Descriptive statistics tables, boxplots}
\usage{
write_boxplot_descStat(
  data,
  fNameMid,
  folder = ".",
  varsExclude = c(),
  varsInclude = NULL,
  varColor = NA,
  statMethod = NULL
)
}
\arguments{
\item{varColor}{variable for colors in boxplots}
}
\description{
2019-10-13 for OBV
}
