% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{brewPalCont}
\alias{brewPalCont}
\title{Brew a color palette from a numerical vector x.}
\usage{
brewPalCont(x, n = 9, name = "OrRd", digits = 2)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Number of colors for RColorBrewer (depends on the name).}

\item{name}{Name of the RColorBrewer palette.}

\item{digits}{Number of digits in x to form categories for colors. May be negative to round to tens, hundreds, etc.}
}
\value{
Named vector of color names of the same length as x.
}
\description{
Uses RColorBrewer palette names, default is OrRd with 9 colors
Use more/less colors (n) for a bigger/smaller range, i.e. mess/less extremes at each end.
Use more/less digits for more/less categories of colors.
}
