% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bayesreg_table.R
\name{create_bayesreg_table}
\alias{create_bayesreg_table}
\title{Creating regression table from \code{brms}}
\usage{
create_bayesreg_table(modellist, modelnames, cirange = 0.95, savedir)
}
\arguments{
\item{modellist}{An \code{list}-object with model objects (from \code{brms}).}

\item{modelnames}{A vector indicating naming of models.}

\item{cirange}{Set the range of credible intervals, \code{Default} is 0.95.}

\item{savedir}{Define path and filename, e.g., "./output/overview.tex".}
}
\value{
Saves a \code{.tex}-file with the regression table.
}
\description{
Creates a regression table of bayesian regression output (from \code{brms})
 into LaTeX with \code{mcmcReg()} from \code{bayesPostEst}.
}
\examples{
# create_bayesreg_table(
#   modellist,
#   modelnames,
#   cirange = .95,
#   savedir
# )

}
