% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create a quarto dashboard}
\usage{
create_dashboard(
  filename = NULL,
  draftname = "_extensions/myhelpers/draft-dashboard.qmd",
  ext_name = "myhelpers"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have
to add \code{.qmd}.}

\item{draftname}{The path to the draft file. If not indicated, the package
included draft file will be used.}

\item{ext_name}{The name of extension to get the files from.
Default 'myhelpers'.}
}
\value{
\code{qmd}-file for a \code{dashboard} based on the template.
}
\description{
Creates personal quarto template for a dashboard in working directory.
}
\examples{
# create_dashboard(
#   filename = NULL,
#   draftname = '_extensions/myhelpers/draft-report.qmd',
#   ext_name = 'myhelpers'
# )

}
