% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a quarto report file}
\usage{
create_report(
  filename = NULL,
  draftname = "_extensions/myhelpers/draft-report.qmd",
  ext_name = "myhelpers",
  path_to_chrome = "/Users/phil/.cache/puppeteer/chrome/mac_arm-119.0.6045.105/"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have
to add \code{.qmd}.}

\item{draftname}{The path to the draft file. If not indicated, the package
included draft file will be used.}

\item{ext_name}{The name of extension to get the files from.
Default 'myhelpers'.}

\item{path_to_chrome}{The path to location of puppeteer chrome distribution.}
}
\value{
\code{qmd}-file for a \code{html}-report based on the template.
 Installs necessary extensions to render the \code{qmd}-file.
}
\description{
Creates personal quarto template for a report in working directory.
}
\examples{
# create_report(
#   filename = NULL,
#   draftname = '_extensions/myhelpers/draft-report.qmd',
#   ext_name = 'myhelpers',
#   path_to_chrome = '/Users/phil/.cache/puppeteer/chrome/mac_arm-119.0.6045.105/'
# )

}
