% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script.R
\name{create_script}
\alias{create_script}
\title{Creates R script template in working directory}
\usage{
create_script(
  filename = NULL,
  kind = "r",
  draftname = "template",
  path_to_templates = "/Users/phil/Documents/templates"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have to
add \code{.R}.}

\item{kind}{indicate if you want to create R (`r`) or Python (`py`) file.}

\item{draftname}{The path to the draft file. Default name is \code{template.R}.}

\item{path_to_templates}{The path to the location of the template, css, and
other files.}
}
\value{
R script based on the template.
}
\description{
This function just creates a R script file based on the saved template as
 draft in the package folder.
}
\examples{
# create_script(
#   filename = NULL,
#   kind = 'r',
#   draftname = 'template',
#   path_to_templates = '/my_path_to_templates'
# )

}
