% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides.R
\name{create_slides}
\alias{create_slides}
\title{Personal quarto template for slides}
\usage{
create_slides(
  filename = NULL,
  draftname = "_extensions/myhelpers/draft-slides.qmd",
  ext_name = "myhelpers",
  path_to_chrome = "/Users/phil/.cache/puppeteer/chrome/mac_arm-119.0.6045.105/"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have to add
\code{.qmd}.}

\item{draftname}{The path to the draft file. If not indicated, the package
included draft file will be used.}

\item{ext_name}{The name of extension to get the files from. Default
"myhelpers".}

\item{path_to_chrome}{The path to location of puppeteer chrome distribution.}
}
\value{
\code{qmd}-Files for slides based on template. Installs necessary
 quarto extensions to render the \code{qmd}-file.
}
\description{
Creating a new quarto file for presentation own defined template.
}
\details{
This function just creates a quarto file based on the saved template as draft
 in the package folder. It installs also the following useful quarto
 extensions:
   - jmbuhr/quarto-qrcode
   - martinomagnifico/quarto-simplemenu
   - mcanouil/quarto-iconify
   - quarto-ext/fontawesome
   - schochastics/academicons
   - quarto-social-embeds
 To print PDF-slides, you need to install \code{decktape} and
 \code{puppeteer} first.
}
\examples{
# create_slides(
#   filename = NULL,
#   draftname = '_extensions/myhelpers/draft-slides.qmd',
#   ext_name = 'myhelpers',
#   path_to_chrome = '/Users/phil/.cache/puppeteer/chrome/mac_arm-119.0.6045.105/'
# )

}
