% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides.R
\name{create_slides}
\alias{create_slides}
\title{Personal quarto template for slides}
\usage{
create_slides(
  filename = NULL,
  path = "./",
  draftname = "draft-slides.qmd",
  path_to_templates = "/Users/phil/Documents/templates/quarto/",
  path_to_chrome = "/Users/phil/.cache/puppeteer/chrome/mac_arm-119.0.6045.105/"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have to add
\code{.qmd}.}

\item{path}{The path were to store and build everything. Default is working
directory}

\item{draftname}{The name of the draft file. Default value is
\code{draft-report.qmd}. With default name it will automatically search for
an \code{assets} folder and copy the complete folder to the project path.}

\item{path_to_templates}{The path to the location of the template, css, and
other files.}

\item{path_to_chrome}{The path to location of puppeteer chrome distribution.}
}
\value{
\code{qmd}-Files for slides based on template. Installs necessary
 quarto extensions to render the \code{qmd}-file.
}
\description{
Creating a new quarto file for presentation own defined template.
}
\details{
This function just creates a quarto file based on the saved template as draft
 in the package folder. If you want full functionality you need to install
 manually the following quarto extensions:
   - jmbuhr/quarto-qrcode
   - martinomagnifico/quarto-simplemenu
   - mcanouil/quarto-iconify
   - quarto-ext/fontawesome
   - schochastics/academicons
   - quarto-social-embeds
 To print PDF-slides, you need to install \code{decktape} and
 \code{puppeteer} first.
}
\examples{
# create_slides(
#   filename = NULL,
#   path = './',
#   draftname = 'draft-slides.qmd',
#   path_to_templates = '/my_path_to_templates/',
#   path_to_chrome = '/Users/phil/.cache/puppeteer/chrome/mac_arm-119.0.6045.105/'
# )

}
