% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides_pdf.R
\name{create_slides_pdf}
\alias{create_slides_pdf}
\title{Creating PDF of slide}
\usage{
create_slides_pdf(
  filename = NULL,
  fragments = FALSE,
  author = "Dr. Philipp Kleer"
)
}
\arguments{
\item{filename}{The name of the created file. You don't have to add
\code{.html}.}

\item{fragments}{Default is FALSE, indicates if you want to print fragments
as single slides.}

\item{author}{Add the author name for the PDF information.}
}
\value{
PDF document of \code{qmd}-slides.
}
\description{
Creating a PDF document out of the \code{qmd}-slides.
}
\details{
This function executes the decktape command in the terminal to create the
 PDF file. Please be aware that decktape and puppeteer needs to be installed.
 Furthermore, you need a folder
 \code{./cache/puppeteer/chrome/mac_arm-119.0.6045.105/} with the correct
 version of chrome to get puppeteer run (might change with updates
 of puppeteer).
}
\examples{
# create_slides_pdf(
#   filename = NULL,
#   fragments = FALSE,
#   author = 'Dr. Philipp Kleer'
# )

}
