% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyptest_to_latex.R
\name{hyptest_to_latex}
\alias{hyptest_to_latex}
\title{Transfer hypothesis' results to TeX}
\usage{
hyptest_to_latex(object, savedir = "./hypothesis.tex", caption = NULL)
}
\arguments{
\item{object}{Object as result of \code{hypothesis()} from \code{brms}.}

\item{savedir}{Define path and filename, e.g., './output/overview.tex'.
Default is \code{'./hypothesis.tex'}.}

\item{caption}{Set the caption for the output. Default is \code{NULL}.}
}
\value{
Prints result of \code{brms::hypothesis()} to LaTeX in console.
}
\description{
Transfers result of \code{brms::hypothesis()} into LaTeX output.
}
\examples{
# hyptest_to_latex(
#   object,
#   savedir = './hypothesis.tex'
#   caption = NULL
# )

}
