% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_beyonce.R
\name{load_beyonce}
\alias{load_beyonce}
\title{Beyonce palette}
\usage{
load_beyonce(pal = 126, textcolor = 5, ticksgrid = 4, legendtext = 1)
}
\arguments{
\item{pal}{Name here the palette you want to utilize in the setting.}

\item{textcolor}{This variable defines textcolor and background.}

\item{ticksgrid}{This color defines ticks and grids.}

\item{legendtext}{This color defines legendtext.}
}
\value{
Set ggplot-Theme to preferred color palette as indicated by the
 function.
}
\description{
Changes ggplot style to specific beyonce style.
}
\examples{
# load_beyonce(
#   pal = 126,
#   textcolor = 5,
#   ticksgrid = 4,
#   legendtext = 1
# )

}
