% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_grey_nogrid.R
\name{load_grey_nogrid}
\alias{load_grey_nogrid}
\alias{load_gray_nogrid}
\title{Colorblind-friendly palette (no grid).}
\usage{
load_grey_nogrid(font = "Fira Sans", color = "white")

load_gray_nogrid(font = "Fira Sans", color = "white")
}
\arguments{
\item{font}{Define the font you want resp. have installed.}

\item{color}{Define the text color.}
}
\value{
A grey styled ggplot-theme without grid.
}
\description{
With this function you alter the ggplot-theme to grey with no grid.
 Furthermore, font is adjusted. If you haven't installed \code{Fira Sans}
 change value. Textcolor can be adjusted by \code{color}
}
\examples{
# load_grey_nogrid(
#   font = 'Fira Sans',
#   color = 'white'
# )

}
