% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pause.R
\name{my_pause}
\alias{my_pause}
\title{Pausing machine.}
\usage{
my_pause(time = 30)
}
\arguments{
\item{time}{Indicate the time you want machine to pause (in seconds).}
}
\value{
Pauses machine and runs garbage collection.
}
\description{
Just a helper function to pause system and then do garbage collection
 (helpful when running a script with intense RAM use).
}
\examples{
# mypause(time = 30)

}
