% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\docType{package}
\name{myHelpers}
\alias{myHelpers}
\title{myHelpers: Misc helper functions for R}
\description{
Misc helper functions for R
}
\section{Base}{
NA
}

\section{STAT + PLOTS}{

}

\section{Write}{

}

\section{Colors}{

}

\section{Conversions}{

}

\section{Plots}{

}

\section{Calculations}{

}

\author{
Peter Juvan, \email{peter.juvan@gmail.com}
}
