% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loo.R
\name{plot_loo}
\alias{plot_loo}
\title{Plots \code{loo}-object after running Bayesian models.}
\usage{
plot_loo(loo, folder = "./", what = "regmodel")
}
\arguments{
\item{loo}{\code{brms} object that has been calculated leave-one-out
criteria with function \code{loo()}.}

\item{folder}{Indicating folder from working directory. Default is \code{./}.}

\item{what}{Indicating model under inspection. Default is \code{regmodel}.}
}
\value{
Saves LOO-diagnostics.
}
\description{
This function creates a diagnostic plot for the \code{loo} calculations
 from a \code{brms} object.
}
\examples{
# plot_loo(
#   loo,
#   folder = './',
#   what = 'regmodel'
# )

}
