% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_to_latex.R
\name{posterior_to_latex}
\alias{posterior_to_latex}
\title{Transfer posterior description to LaTeX}
\usage{
posterior_to_latex(
  descobject,
  digits = 3,
  minimal = TRUE,
  savedir = "./posterior.tex"
)
}
\arguments{
\item{descobject}{Object as result of \code{describe_posterior()} from
\code{bayestestR}.}

\item{digits}{Set the digits that should be shown in the latex output.}

\item{minimal}{Shrinks output to relevant parts. Excludes ROPE definition,
PS, and Rhat. Default is \code{TRUE}.}

\item{savedir}{savedir Define path and filename, e.g., './output/overview.tex'.
Default is \code{'./posterior.tex'}.}
}
\value{
Returns LaTeX output of the posterior description by
 \code{bayestestR::describe_posterior()}
}
\description{
This function transfers results from function
 \code{bayestestR::describe_posterior()} into latex output
 (with \code{xtable}).
}
\examples{
# posterior_to_latex(
#   descobject,
#   digits = 3,
#   minimal = TRUE,
#   savedir = './posterior.tex'
# )

}
