% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scores.R
\name{read_scores}
\alias{read_scores}
\title{read_scores}
\usage{
read_scores(models = "ecmwf-system4", scores = c("EnsCorr", "FairRpss"),
  indexes = "tas", methods = "none", initmonths = "05",
  grids = "global2", granularities = "seasonal", obs = NA, periods = NA,
  seasmethods = "none", detrends = FALSE, leads = 2, reference = NULL,
  dpath = "/store/msclim/bhendj/EUPORIAS", cleanup = FALSE)
}
\arguments{
\item{models}{character vector of the names of forecasting systems}

\item{scores}{character vector of skill scores to be read}

\item{indexes}{character vector of variable names}

\item{methods}{character vector of bias correction methods}

\item{initmonths}{character vector of initialization months}

\item{grids}{character vector of grids}

\item{granularities}{character vector of temporal resolutions}

\item{obs}{character vector of verifying observations (e.g. ERA-INT)}

\item{periods}{character of periods over which the verification has been carried out (e.g. 1981-2014)}

\item{seasmethods}{logical should recalibrated indices be used?}

\item{detrends}{logical should detrended series be used?}

\item{leads}{index of lead times, or a character string specifying
what lead times to use (currently only \code{"last"} is supported)}

\item{reference}{name of reference model against which the forecast skill
is evaluated. If set to \code{NULL} (the default), forecasts are
verified against a climatological forecast.}

\item{dpath}{head of directory tree}

\item{cleanup}{logical, should only non-missing combinations be retained?}
}
\description{
Read skill scores to data frame
}
\keyword{util}

