% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_scores.R
\name{read_scores}
\alias{read_scores}
\title{read_scores}
\usage{
read_scores(models = "ecmwf-system4", scores = c("EnsCorr", "FairRpss"),
  indexes = "tas", methods = "none", initmonths = "05",
  grids = "global2", granularities = "seasonal", ccrs = c(FALSE, TRUE),
  detrends = FALSE, leads = 2, reference = NULL,
  dpath = "/store/msclim/bhendj/EUPORIAS")
}
\arguments{
\item{models}{character vector of the names of forecasting systems}

\item{scores}{character vector of skill scores to be read}

\item{indexes}{character vector of variable names}

\item{methods}{character vector of bias correction methods}

\item{initmonths}{character vector of initialization months}

\item{grids}{character vector of grids}

\item{granularities}{character vector of temporal resolutions}

\item{ccrs}{logical should recalibrated indices be used?}

\item{detrends}{logical should detrended series be used?}

\item{leads}{index of lead times, or a character string specifying
what lead times to use (currently only \code{"last"} is supported)}

\item{reference}{name of reference model against which the forecast skill
is evaluated. If set to \code{NULL} (the default), forecasts are
verified against a climatological forecast.}

\item{dpath}{head of directory tree}
}
\description{
Read skill scores to data frame
}
\keyword{util}

